Create LICENSE
I'm proposing an MIT license for review.
 Needless to say, as my work has been so far limited to usability/man and distribution channel additions, 
the addition of my name to the license is simple license protocol but in no way entitles me to said addition, i.e. it is the developer doing the actual code, Sanel, that decides on who is on the copyright :)

I'm doing this as a separate branch to facilitate license review separately from the code.